Parser class: Edit getContent method
getContent method checks for missing content by splitting the user input with a space, before checking
for the length of the array and indexing the array at the second element to check the string.

getContent method does not check for missing content correctly. It incorrectly checks the split command
by array length and empty content in the same if condition.

Let's separate the two conditions into two separate if conditions instead of putting it together.

By separating the two checks, it accurately checks the array length first before trying to index the array.