Add a way to attach priority to tasks
Users currently are not able to sort and tag their tasks based on
its priority, which may reduce user experience.

Let's,
* add a feature to allow users to state the priority of their tasks
using the /priority command attached with each task
* add a feature where users can find all lists of a particular priority
using the "priority" command

This allows users to attach a priority to each task and quickly
obtain a list of all tasks of a certain priority.