Include assertions
Add assertion to ensure vital assumptions hold in operations.

The various commands in has assertions to ensure there is always
the adequate number of details in every case.

Read from save would always check and create a save file if not
found before the operation of the chatbot to ensure there would be
a save file to write to.

Delete function would check to ensure access to task is bounded by
the size so no deletes can occur with index larger than the size of
task list.