Improve code quality.
There were many unneccesary and unoptimised code present in the
last commit due to the drastic change of abstracting command in
the readInput method of the Parser class.

Implemented more inline variable, removed unused imports, removed
the unused fields in Ui class and minimise scope of variables,
specifically in the Ui class.

This is to improve the code quality, according to W5.4