Add assertions to code.
There are several methods that require instances of classes to be passed
in but in the off chance that they are null, the application will not work as
intended.

Add assertions for those methods to ensure class instances are not null.

Using assertions is preferable over exceptions as it is not a user error
if the instances passed in are null.