Add assert feature
Assertions help catch bugs during the development phase by checking assumptions that the code relies on. It also makes the code easier to understand and maintain.
Reasons for adding assertion in the Repsmax, Storage, Task and TaskList Class:
Repsmax: helps to identify empty file path and user input
Storage: helps to identify empty file path, failure of file path and existence of the file
Task: helps to identify empty descriptions
TaskList: ensures that task input is not a null