Update javadocs
Currently no javadocs are added for the new methods created after the code was refactored.

The lack of javadocs may make it unclear to the reader the purpose and function of each method.

Adding javadocs makes it easier for reader to understand the code logic as it improves the readability of the code.

Let's add javadocs for the purpose, parameters and return values for all new methods created.

Adding javadocs makes it easier to understand all existing code.