Improve code quality
Some of the methods in the code were too long (longer than 30 lines of
code) which affects the readability of the code.

Some boolean fields were also not named properly which may
cause confusion.

Broke down the long methods into smaller and shorter methods
such that it is easier to read.

Change the name of the boolean fields.