Add Undo extension
Currently no undo function call for Duck.

This makes it inconvenient for the user to reverse their last command, as they will have to manually reverse it themselves.

Adding an undo extension makes it easier for users to undo their previous command, in cases where they made a mistake or change their mind.

Let's add an Undo extension by using a new class Undo, to track the previous command and reverse it if called.

Using a new class to handle this undo feature prevents the logic from getting confusing and messy, if it were all stored in the Parser class instead.