Add some assertions to ensure the program run properly
There are possible errors which trigger null value or wrong
LocalDateTime value in the program.

Let's
* Add assertion in execute(TaskList, Storage) in Command class
to make sure the TaskList and Storage passing to the method is
not null.
* Add assertion to the constructors in Task class and its
subclasses to make sure the String passing to the constructor is
 not null
* Add assertion to the constructors of Event class to make sure
the start value is before the end value
* Add assertion to the mark(int), delete(int), unMark(int) to make
sure the task in the given position (int) is not null
* And some other small assertions