Update chatbot to list tasks according to priority
Tasks are listed according to when they are added.

To better improve the user experience, tasks should be grouped by
priority with HIGH priority being printed first.

Let's separate the printing of tasks based on priority.

This also allows the code to be more flexible. It is easy to update the
chatbot if new priorities are added.