Refine CI using GitHub Actions
The project fails automated testing and CI due to
last three statements in gradle.yml file.

The last three segments of the workflow related to
I/O redirection tasks were removed, as they are not
applicable to this project.

This CI setup ensures that all commits are automatically
verified. GitHub Actions was chosen for its seamless
integration with GitHub repositories.