Add assertions for Storage and Nameless class
Used assertions in the Nameless class to
ensure that the inputs are not null in the
todo, deadline and event methods. The assertions
will help to ensure that users cannot input null.

Used assertions in the Storage class to verify
the creation of directory and file in the
createDirectory and createFile methods. The
assert statements will ensure that file and
directory exists after program execution.