fix: Change parameters of extracted function
Matcher groups separates the string according to the
regex pattern specified.

During refactoring and extraction of function it is crucial to
ensure the correct group was passed into the handling function.

Passing in the correct group of matched strings allows
the extracted function to perform the expected behaviour as if
the code was written sequentially instead of reprocessing the
entire string.

Let's check that Matcher groups are properly passed into the
extracted functions especially in Parser and Bee where
extracted functions are heavily utilized.