Move the error catching from the Bopes class to the Parser class
Currently error handling is done directly in the Bopes class and all
its component propogate the error message up to the Bopes class.
This is not ideal as it makes the code harder to read and maintain.
It would be better to move the error handling to the Parser class
as it is responsible for parsing the user input.

This will help to make the code more modular and easier to read.
The Parser class is responsible for parsing the user input,
so it makes sense for it to handle the error catching as well.

With this change, the Bopes class will only need to call the Parser
and it will also be easier to debug.