Update Parser.java
Delete command: prints the wrong output
Mark command: prints the wrong output

Bote command should output the command has been executed.

The method returns the String to be output through the method that is being called in TaskList.

This allows the Parser command to print the correct output instead of the 'Wrong Format' message.