Extract methods from parse method to make it shorter
Previously, the parse method had the if-else conditionals for each command as well as the logic to execute for each command.
This made the method very long with more than 30 lines of code.

It would be better if we abstracted out some of the logic into methods to make the method shorter.

Let's abstract out the logic for each command word into their individual function, so that we can
reduce the number of lines of code for the parse method.
 Please enter the commit message for your changes. Lines starting