Add github actions initial .yml file
Previously, there was no CI pipeline and we need a human to run the automated test cases and verify that everything is working.

However, it would be better if github ran test cases automatically after we push a new version, and this would help us in assessing whether a piece of
code is correct.

Let's add a github actions config .yml file to test this.