Update TaskList.java Update Ui.java
TaskList class: the get function assume that the i that is passed in is positive and less than the size of task list
Ui class: the printList() function assume that the TaskList.length() function will return a positive number

Assertion is added to both of these methods to make sure the assumption holds.

This is to make sure that the code does not fail due to the carelessness of the programmer.