Update Java Doc for TaskEditor.java
The javadoc is based on the old method before refactoring the code.

The doc explains the older format of the method when it dealt with
all of marking, unmarking and deletion. This is confusing for any
new developers of the project.

Let's update the Javadoc such that it reflects the new functions of
the method after refactoring.

This makes the method more understandable for new developers when
they need to modify the method or use the method for other methods
of the program.