Modify TaskListStorageTest.java
The test is not setting up and cleaning the relevant files before
and after the test.

This leads to issues in passing the test especially if the storage
list file already contained some items in it.

Lets add "@BeforeEach" and "@AfterEach" command for this file to
set up proper testing environments before and after each test case
while also following the DRY (Don't Repeat Yourself) principle.

This allows the testing process to be smoother for further builds
and github continuous integrations.