Update AI.md
The AI-assistance documentation has not been updated in a while.

A lot of AI assistance was used for many methods and is difficult to
keep track of.

Lets update the AI.md to quickly document all the ways AI was used
and lessons learnt during the journey.

This will help close learning loop and give credit where its due
to prevent accusations of plagiarism.