Update Event.java
There is a regression error that is spotted on Event.java due to
added space before the "/from" inside of the getDescription() method.

This causes one of the test cases to fail which prevents continuous
integration inside github.

Let's fix the regression error inside Event.java.

This will allow the build to be created successfully and CI to be
smooth.