Add a way for Priority to be stored in each task
We are lacking a way to add priority for each task, but it would be good
for the user to indicate the priority for each of the tasks that he has.
However, in the case that the user does not want to add a priority to
the task, he would also be allowed to do so.

Let's create a Priority class that encapsulates the priority of each
task, and update the Task class and its subclasses to be able to store
and display the priority of each individual task when each command is
executed.