Update TimedTask.java
The locale of the DateTimeFormat class was not specified.

When different OS or machines use the program, the program will
follow their specific device locale and this will cause problems for
the storage of the times.

Let's add a specific locale to follow for the program.

This will allow the behaviour of the DateTimeFormat class to be
consistent across many different devices and OS.