Make the application exit upon error loading data file
If there is an error creating the data file or the data
file is not of the correct format, the chatbot proceeds
to run as usual.

Having the chatbot run even without a proper data file
can introduce bugs during runtime and worsen user
experience.

Let's make the chatbot verify the data file upon
initialization and exit if there is no proper data
file.

This allows the user to know that there is an issue
regarding the data file and should check their hard
disk or delete the old data file instead of encountering
bugs when using the chatbot.