TaskList.java: Implement Stream
The current implementation of find method is not very straightforward
because it involves a lot of logics like Regex pattern and Matcher which
makes the code looks messy.

Change the iterative implementation to stream as it offers a more
straightforward pipeline in filtering out the desired tasks that matches
the given pattern. Hence improves the readability of the code