Update Parser.java Update Storage.java Update Task.java Update TaskList.java Update Ui.java
The current implementation does not have a priority feature.

A priority feature will be useful for users to set tasks that are more urgent at the top of the list

The user can now set to HIGH, MEDIUM or LOW for each tasks with LOW being the default.

It is done this way so tasks with HIGH priority will be shown at the top.