Add git message and explain Level-10
The previous version of Pebble was interacted through the Command Line
Interface.

This is not very user friendly as it requires users to know the basics
of running the code in a code editor.

A Graphic User Interface (GUI) is added to provide users another way
of interacting with Pebble, using JavaFX, FXML and CSS. This way a
fatJar file can be created to encapsulate the code into an executable,
making it even more user friendly.