Implement basic JavaFX UI
Copy code from the JavaFX tutorial to implement a basic, but working version of a Chat Bot UI. No configuration is changed except that necessary to have it work with Torne.

Modify Torne such that it still can run as a CLI, but ChatOutput also stores the last Torne output so a GUI can fetch it and display it as a message. This also means that, as the GUI is running, Torne also prints the equivalent commands to the terminal.