Add BCD extension
Currently, users can only add new tasks but cannot modify existing ones
without deleting them.

A more flexible task feature is required to let users modify the
existing tasks.

We add an extension to introduce the ability to modify the deadline
associated with the deadline in Deadline tasks and time frame
for Event tasks by adding an edit task method to the Task class.

This is so that OOP principles are adhered to still.