Improve coverage of assertions in code
Previously, assertions mainly spanned in the task list as a means of out
of bounds checking.

This commit improves upon the power of assertions by enforcing certain
requirements before storage in text form is read and processed by the
program. Doing so prevents the damage that corrupted text files could
cause to the normal function of the program.