Modify test cases
Current test cases do not work due to changes to the Event and Deadline
classes when implementing the extension

The test cases for Event need to change to be appropriate in verifying
the Event behaviour

Change the Parser stub so that it now produces the desired behaviour
when used by the Event class. Modify the test cases to not throw
errors anymore.

This is the best way to get the test cases to work properly