Command.java: Change class to an abstract one
All Common Subclasses implement the getResponse() method.

Hence, it makes sense for the Command Class to be abstract, containing an abstract method getResponse().

Furthermore, although some Subclasses do not require for their getResponse() methods to return anything,
I have made it so that they will artificially return a TaskTracker object, to fit into the constraints of the abstract method.