Modify TaskListCreatorTest.java
There is repetition within the test methods of
TaskListCreatorTest.java.

This reduces the code quality of the testing phase and makes new
methods more difficult to test as the setup will have to be repeated
again.

Let's extract out the setup from each method into a class level
implementation.

This reduces the amount of repetition and allows us to change the
test cases as quickly as possible. This also causes the test to test
the behaviour of each class rather than the execution only.