Refactor code to improve code quality
Most of the code for Storage.java has a lot of arrowhead style code and
nesting.

The method loadTasks() is also really long leading to a lot of LoC.

By extracting out the behaviour of a few lines of code in loadTasks(),
we can reduce the margin for error and helps debugging the code easier
as well.

Let's extract out a few behaviours in loadTasks() to other methods to
make this easier to read.