Add JavaDoc comments to Storage class
There are methods in the Storage class that is not documented with
JavaDoc comments.

Adding JavaDoc comments would improve the documentation.

As a step towards better documentation, let's add JavaDoc comments to
the Storage class. Doing so will make subsequent function calls easier
as there are JavaDoc comments to refer to.