Fix bug where not all lines from text file are read
Only one line of the text file is read and it is read multiple times.
This leads to inaccurate ArrayList of Tasks extracted from the file.
There is no new line character added to the text lines, causing the
task sentences to be combined in a single line in the text file.

Let's remove the duplicate readLine() method of bufferedReader in the
load() method of Storage and add "\n" to the task sentences.

Also fix coding quality issues including indentation, white spaces
and refactoring startDate and endDate to two separate methods in
Storage class.