Add JavaDoc comments to Deadline class
There are methods in the Deadline class that is not documented with
JavaDoc comments.

Adding JavaDoc comments would improve the documentation.

As a step towards better documentation, let's add JavaDoc comments to
the Deadline class. Doing so will make subsequent function calls easier
as there are JavaDoc comments to refer to.