Add JavaDoc comments to Todo class
There are methods in the Todo class that is not documented with
JavaDoc comments.

Adding JavaDoc comments would improve the documentation.

As a step towards better documentation, let's add JavaDoc comments to
the Todo class. Doing so will make subsequent function calls easier
as there are JavaDoc comments to refer to.