Implement SLAP in Storage class.
Instead of processing the input within the loadTask method, the processing is abstracted into individual methods.

The additional methods abstract out the processing of the string into its relevant components so that the loadTask method's logic is mainly to add the tasks to the tasklist.
This makes it easier to understand the logic of the loadTask method without knowing how it is implemented.