Improve Code Quality
There are several instances of poor variable names and lack of comments on the WHAT & WHY aspects of the code.

Following the Code Quality Convention would greatly improve the code quality and readability of the code.

As a step towards code quality, let's ensure that code does not violate any of the Code Quality Conventions, such as repeated code, bad variable names, lack of comments and wrong code style.