Integrate assert testing within Shnoop.java
There are several classes pivotal to the functioning of the ChatBot.
If any of the classes do not function as intended, debugging would
be a difficult and complicated process.

Performing assumption checks at the start of the code would be most
useful.

As a step towards fool-proofing the application, let's perform assert
tests on one of the main classes that is invoked and initialised
from the very beginning. Doing so will check to see if any assumptions
do not hold true at the very base and beginning invocation of the
program.