Remove static strings in ui
There were several Strings that were cached in Ui, however it was not
consistent with all the strings.

Changing it to have all Strings initialised during method calls makes
it more consistent and improves code quality.

In a bid to improve consistency, let's move all String so that they
are only initialised when the method is called rather than it being
cached on object creation.