Assert Feature: Add Assert Feature to different classes to use for debugging.
Assert statements help with identifying bugs related to assumptions
made about certain values or actions. This will be useful when
invalid input entries are given by users which could result in the
code not running properly.

Let's check arguments that are using user inputs and task list
as arguments to verify if any invalid inputs or empty lists are
passed as values.

Checking these values will be essential as these are the common
values and arguments used in the code to execute other methods
and commands for user to receive output.