Add User Guide to README.md and Ui.png
Let's add user guide to readme which covers all the basic functionalities, their command,
sample input and output. This is to allow user to have a guide on how to run the bot and
the different commands that they can use in this bot. Ui.png is to show how the interface
looks like for more detail.