Prevent duplicate tasks from being added Previously, duplicate tasks could be easily added into the task list, which is undesirable as the user may be confused about why multiple tasks with the same description are present.
A duplicate-checking method has been added to scan and ensure that no
task descriptions match before a task can be added. If a duplicate task
is added, addition of the task is aborted and the user is informed.

Streams are used as part of this feature to convey the stateless nature
of the duplicate checker method, allowing for multi-threaded use if
necessary.