Refactored code to follow SLAP.
There was code within Nave class that handled the core functionality of the bot, calling the parser, ui and task storage.

The code that handled getting the user input and parsing it, subsequently outputting a response has been abstracted out.

The code for the function run(), now better follows the SLAP principle.

In addition, duplicated code for parsing the user input for the GUI vs the command line has been consolidated to prevent code duplication.